### 2024-09-23 #7269 @editorialbot
[PRE REVIEW]: cdsaxs: A model fitting package for CD-SAXS data analysis
**Submitting author:** @mrhxszo (<a href="http://orcid.org/0009-0007-8982-424X">nischal dhungana</a>)
**Repository:** https://github.com/CEA-MetroCarac/cdsaxs
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.0.2
