+++ CONTRIBUTING.md
@@ -1,0 +1,21 @@
+# Contributing to cdsaxs
+
+cdsaxs welcomes contributions from everyone! Here's how you can get involved:
+
+1. (Optional) Have an idea for cdsaxs? Share it! Create a new GitHub Issue detailing what you'd like to do. This helps us discuss and refine your contribution before code is written.
+2. Fork cdsaxs and make your changes. Not sure how? This GitHub guide [make your changes](https://github.com/firstcontributions/first-contributions) can help beginners.
+3. Once you're happy with your changes, open a Pull Request on GitHub. This lets us review your contribution and discuss any improvements.
+4. Make sure all tests pass! This ensures your changes don't break existing functionality.
+5. Once everyone's happy, we'll merge your code into cdsaxs and make a new release.
+
+For more detailed contribution guidelines, refer to our official Contributing Guide: [Contributing Guide](https://cea-metrocarac.github.io/cdsaxs/for_developpers.html).
+
+## Reporting Bugs
+
+Found a bug in cdsaxs? Help us fix it! Open a new GitHub Issue detailing the problem you encountered. Here are some tips for good bug reports:
+
+* Make sure the issue can be reproduced consistently.
+* Include details like the package versions you're using and your operating system.
+* The more information you provide, the easier it is for us to diagnose and fix the bug.
+
+By following these guidelines, you can help make cdsaxs even better!
