### 2024-09-24 05e1f4b
minor improvements for contributing
