### 2024-09-24 e9fea37
joss logo added
