### 2024-09-26 #7269 @mrhxszo
@Kevin-Mattheus-Moerman Thank you for your help. I've added `CONTRIBUTING.md` and linked it on my `README.md`. Please let me know if you see further room for improvements.
