### 2024-10-18 #7269 @srmnitc
@ZhenchenHong @benjaminbolling @larsenkg @trygvrad, would any of you have time and be willing to review this submission for JOSS? As you might know, we carry out our checklist-driven reviews here in GitHub issues and follow these [guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html). Thanks for your time!
