### 2024-12-17 704121d
symmetric case improved
