### 2024-12-18 09a3a87
"test modified for symmetric case"
