### 2024-12-19 d2cfef2
error for symmetric case strong_castle fixed
