### 2025-01-06 CEA-MetroCarac/cdsaxs#1 @benjaminbolling
JOSS review: Dependencies
The list of dependencies should be stated an environment/requirements files. It is not clear what the docs_requirements.txt infers.
