### 2025-01-06 CEA-MetroCarac/cdsaxs#2 @benjaminbolling
JOSS review: Hard-coded values in fitter.py
The do_stats method in fitter.py uses hard-coded values for cf to define the z-value: "This method generates a set of statistical data on the best fit parameters obtained from the MCMC fitting process". These values should be available through an input file and not hardcoded, and use ranges rather than explicit values. (e.g. cf==0.94 -> z(cf==0.99) > z(cf==0.95)).
