### 2025-01-06 CEA-MetroCarac/cdsaxs#3 @benjaminbolling
JOSS review: Runtime warnings
_cdsaxs/fitter.py:289: RuntimeWarning: invalid value encountered in scalar divide_ when running Tutorials/strong_castle_simulation.ipynb

_site-packages/emcee/autocorr.py:38: RuntimeWarning: invalid value encountered in divide acf /= acf[0]_ when running Tutorials/stacked_trapezoid_simulation_and_fitting.ipynb
