### 2025-01-06 CEA-MetroCarac/cdsaxs#4 @benjaminbolling
JOSS review: Random seed range in mcmc_bestfit_stats
Why was the random seed range (if none is provided) in mcmc_bestfit_stats selected to be 2 ** 32? Same goes for burnin steps, of which both may be beneficial to allow user of the software to define manually as an input argument.

Otherwise, add a comment why these values were selected.
