### 2025-01-06 CEA-MetroCarac/cdsaxs#5 @benjaminbolling
JOSS review: Statement of Need, last paragraph
Would suggest rephrasing "democratizes access" (last paragraph of the Statement of Need section).
