### 2025-01-06 CEA-MetroCarac/cdsaxs#6 @benjaminbolling
JOSS review: Previous work
The paper states that existing tools are either proprietary or insufficiently flexible. It would be useful to mention examples of such and why they are deemed either proprietary or insufficiently flexible, and the state of those.

Would thereafter also recommend referring to previous work in the field of data analysis software or methods for CD-SAXS.
