### 2025-01-06 CEA-MetroCarac/cdsaxs#7 @benjaminbolling
JOSS review: Benchmarking
It would be useful to benchmark the output of the software in comparison to the data published in the field of CD-SAXS from others' scripts/software.
