### 2025-01-06 #7559 @benjaminbolling
First review done, sorry for the delay. Very nice software package!

---
### Functionality
- Functionality will be resolved with:
  - [!2](https://github.com/CEA-MetroCarac/cdsaxs/issues/2)

### Documentation
- Installation instructions will be resolved with [!1](https://github.com/CEA-MetroCarac/cdsaxs/issues/1)
- Functionality documentation will be resolved with [!4](https://github.com/CEA-MetroCarac/cdsaxs/issues/4)

### Software paper
- _State of the field_  and _References_ will be resolved with
  - Comparison to other software/scripts [!6](https://github.com/CEA-MetroCarac/cdsaxs/issues/6)
  - Benchmarking [!7](https://github.com/CEA-MetroCarac/cdsaxs/issues/7)

---

## Recommendations not required for the paper
I leave these to the author for consideration:
- Runtime warnings, [!3](https://github.com/CEA-MetroCarac/cdsaxs/issues/3)
- Rephrasing in "Statement of Need", [!5](https://github.com/CEA-MetroCarac/cdsaxs/issues/5)

---

I may have further comments on functionality once I get more time to run some tests.
