### 2025-01-06 CEA-MetroCarac/cdsaxs#2 @benjaminbolling
Requirement for http://openjournals/joss-reviews#7559.
