### 2025-01-06 CEA-MetroCarac/cdsaxs#7 @benjaminbolling
Requirement for openjournals/joss-reviews#7559.
