### 2025-01-06 #7559 @benjaminbolling
@marcocamma - I have noticed that the checklist is not generated unless you manually write/tag the editorialbot
