### 2025-01-06 CEA-MetroCarac/cdsaxs#8 @marcocamma
JOSS Review: installation
Installation from github fails (on ubuntu 22.04 machine), see terminal output at the en

Also, it would be nice do suggest to create a venv environment
```bash
python -m venv /path/to/new/virtual/environment
source /path/to/new/virtual/environment/bin/activate
```

```bash
cammarat@lapcammarata:~/linux$ pip install git+https://github.com/CEA-MetroCarac/cdsaxs.git
Defaulting to user installation because normal site-packages is not writeable
Collecting git+https://github.com/CEA-MetroCarac/cdsaxs.git
  Cloning https://github.com/CEA-MetroCarac/cdsaxs.git to /tmp/pip-req-build-pztj11e0
  Running command git clone --filter=blob:none --quiet https://github.com/CEA-MetroCarac/cdsaxs.git /tmp/pip-req-build-pztj11e0
  Resolved https://github.com/CEA-MetroCarac/cdsaxs.git to commit d2cfef280e9cf21a3451ba61a41b9ea8e1271f93
  Installing build dependencies ... done
  Preparing metadata (pyproject.toml) ... error
  error: subprocess-exited-with-error

  × Preparing metadata (pyproject.toml) did not run successfully.
  │ exit code: 1
  ╰─> [14 lines of output]
      Traceback (most recent call last):
        File "/usr/lib/python3/dist-packages/pip/_vendor/pep517/in_process/_in_process.py", line 363, in <module>
          main()
        File "/usr/lib/python3/dist-packages/pip/_vendor/pep517/in_process/_in_process.py", line 345, in main
          json_out['return_val'] = hook(**hook_input['kwargs'])
        File "/usr/lib/python3/dist-packages/pip/_vendor/pep517/in_process/_in_process.py", line 164, in prepare_metadata_for_build_wheel
          return hook(metadata_directory, config_settings)
        File "/tmp/pip-build-env-hz6ksjiz/overlay/local/lib/python3.10/dist-packages/hatchling/build.py", line 117, in prepare_metadata_for_build_wheel
          f.write(builder.config.core_metadata_constructor(builder.metadata))
        File "/tmp/pip-build-env-hz6ksjiz/overlay/local/lib/python3.10/dist-packages/hatchling/metadata/spec.py", line 546, in construct_metadata_file_2_4
          if metadata.core.license:
        File "/tmp/pip-build-env-hz6ksjiz/overlay/local/lib/python3.10/dist-packages/hatchling/metadata/core.py", line 677, in license
          from packaging.licenses import canonicalize_license_expression
      ModuleNotFoundError: No module named 'packaging.licenses'
      [end of output]

error: metadata-generation-failed

× Encountered error while generating package metadata.
╰─> See above for output.

note: This is an issue with the package mentioned above, not pip.
hint: See above for details.
```
