### 2025-01-17 0f612ff
updated readme to include virtualenv installation instruction
