### 2025-01-17 ffacdca
build docs command for ci.yml
