### 2025-01-17 CEA-MetroCarac/cdsaxs#1 @mrhxszo
In accordance with [PEP 725](https://peps.python.org/pep-0725/), the list of dependencies is declared under the `dependencies` key in the `pyproject.toml` file which is used by [Hatch](https://hatch.pypa.io/latest/) to build the project following the [PEP 518](https://peps.python.org/pep-0518/) standard. Similarly, optional dependencies are specified under `project.optional-dependencies`.

```toml
dependencies = [
  "numpy",
  "pandas",
  "emcee",
  "deap",
  "scipy",
  "corner",
]

[project.optional-dependencies]
gpu = ["cupy>=9.0.0"]
```

Given this structure, I chose not to create environment or requirements files, as these were primarily used in older dependency management practices.

However, I agree that placing `docs_requirements.txt` (intended to manage documentation build dependencies) at the root of the project may lead to confusion. To address this ambiguity, I have moved it to the `docs/` directory for better clarity and organization.
