### 2025-01-17 CEA-MetroCarac/cdsaxs#2 @mrhxszo
I've updated the do_stats function to calculate the z value using a function instead of hardcoded values:

```python
        z = stats.norm.ppf(1 - (1 - cf) / 2)
```
