### 2025-01-17 CEA-MetroCarac/cdsaxs#8 @matbryan52
Thank you @marcocamma for the time spent reviewing.  I definitely agree that encouraging a `venv` would be a good addition to the documentation.

However, I believe the installation issue you are having is out-of-scope for `cdsaxs`, and is due to the test install using a system python with a version of `pip` which doesn't quite properly isolate its build environment. I was able to reproduce the issue on a VM with system-packaged `python==3.10.12` and `pip==22.0.2`, but creating a `venv` or otherwise upgrading the system pip to version 22.3.0 or later allows this package to install normally. I wouldn't be surprised if this issue (fixed in this PR https://github.com/pypa/pip/issues/6264 in pip 22.3.0) is related.
