### 2025-01-20 CEA-MetroCarac/cdsaxs#4 @mrhxszo
I've updated the code for calculating burn-in as follows:

```python
 #Autocorelation time to find burnin steps, if not provided by user and emcee fails to do so 1/3 of the
 #first nsteps are discarded. 1/3 period is the common heuristic for MCMC burnin
try:
    if tau is None:
        tau = Sampler.get_autocorr_time(tol=5)
    burnin = int(2 * np.max(tau))
except (emcee.autocorr.AutocorrError, ValueError) as e:
    print(f"Warning: Failed to compute autocorrelation time. Using default burn-in which is 1/3 of nsteps. Error: {e}")
    burnin = int(nsteps / 3)
```
This implementation prioritizes a user-provided `tau` if available. Otherwise, it attempts to compute `tau` using the `emcee` library. Since `emcee` may struggle with certain datasets, a fallback default of 1/3 of `nsteps ` is applied, and a warning is logged to inform the user.

Similarly, here's the modification i made for seed:
```python
# Assign the seed given to user if the format is correct
if seed is not None:
    if isinstance(seed, int) and seed >= 0:
        np.random.default_rng(seed)
    else:
        raise ValueError("Seed must be a non-negative integer.")
```
I've used the `numpy`'s updated way of assigning seed and removed the `2 ** 32`.
