### 2025-01-22 e50af27
Added references to new articles in bibliography and update paper description for clarity
