### 2025-01-22 a4b7234
Update paper.md to answer previous work issue
