+++ joss/paper.md
@@ -33,7 +33,7 @@ CD-SAXS (Critical Dimension Small Angle X-ray Scattering) is an emerging and pro
 
 # Statement of need
 
-CD-SAXS is a powerful yet emerging technique for characterization of nano-components in semiconductor industry, but its potential is currently limited by the lack of open-source software for comprehensive data analysis. Existing tools are either proprietary or insufficiently flexible, leaving researchers with the challenge of developing their own solutions for simulating and fitting CD-SAXS data. Moreover, the diversity of samples analyzed using CD-SAXS requires versatile software that can accommodate different types of models and experimental conditions.
+CD-SAXS is a powerful yet emerging technique for characterization of nano-components in semiconductor industry. Efforts by the community were done to explore different combination of algorithms to model CD-SAXS data [@hannon2016advancing, @sunday2016mcmc]. However, there is a lack of open-source software for comprehensive data analysis. It currently on exist a python package inside **Xi-cam**[@Xi-cam] repository but the code is no longer maintained. While it exists proprietary tools, developement of model for CD-SAXS requires researchers to develop their own solutions for simulating and fitting CD-SAXS data. Moreover, the diversity of samples analyzed using CD-SAXS requires versatile software that can accommodate different types of models and experimental conditions.
 
 The cdsaxs package is designed to address this critical gap by providing a modular, open-source solution tailored for CD-SAXS data analysis. It includes two robust models for simulating CD-SAXS data, while also allowing researchers to integrate their own models. This flexibility is crucial for testing and validating models against experimental data, making the development process more streamlined and accessible.
 
