### 2025-01-23 4a9169b
empirical factor c added in mcmc and paper modified for previous work
