+++ joss/paper.md
@@ -34,11 +34,11 @@ CD-SAXS (Critical Dimension Small Angle X-ray Scattering) is an emerging and pro
 # Statement of need
 
 CD-SAXS is a powerful yet emerging technique for characterization of nano-components in semiconductor industry. Efforts by the community were done to explore different combination of algorithms to model CD-SAXS data [@hannon2016advancing, @sunday2016mcmc]. However, there is a lack of open-source software for comprehensive data analysis. **Xi-cam**[@Xi-cam] is the only publicly available python package that contains CD-SAXS data analysis but the code is no longer maintained. While there are few other proprietary
-softwares we don't have access to them. Thus, development of model and it's analysis for CD-SAXS requires researchers to develop their own solutions for simulating and fitting. Moreover, the diversity of samples analyzed using CD-SAXS requires versatile software that can accommodate different types of models and experimental conditions.
+softwares, we don't have access to them. Thus, development of model and it's analysis for CD-SAXS requires researchers to develop their own solutions for simulating and fitting. Moreover, the diversity of samples analyzed using CD-SAXS requires versatile software that can accommodate different types of models and experimental conditions.
 
 The cdsaxs package is designed to address this critical gap by providing a modular, open-source solution tailored for CD-SAXS data analysis. It includes two robust models for simulating CD-SAXS data, while also allowing researchers to integrate their own models. This flexibility is crucial for testing and validating models against experimental data, making the development process more streamlined and accessible.
 
-A key feature of this package is its separation of the simulation and fitting processes, enabling users to concentrate on model development and data analysis without being encumbered by technical complexities. The package is optimized for performance, with support for parallelized fitting on both CPUs and GPUs, significantly enhancing the speed and efficiency of data processing. The fitting process in cdsaxs is powered by the CMAES (Covariance Matrix Adaptation Evolutionary Strategy) algorithm, known for its rapid convergence for x-ray fitting[@hannon2016advancing]. This efficiency allows for real-time data fitting during experiments, empowering researchers to dynamically adjust experimental parameters based on immediate feedback from the analysis.
+A key feature of this package is its separation of the simulation and fitting processes, enabling users to concentrate on model development and data analysis without being encumbered by technical complexities. The package is optimized for performance, with support for vectorised fitting on both CPUs and GPUs, significantly enhancing the speed and efficiency of data processing. The fitting process in cdsaxs is powered by the CMAES (Covariance Matrix Adaptation Evolutionary Strategy) algorithm, known for its rapid convergence for x-ray fitting[@hannon2016advancing]. This efficiency allows for real-time data fitting during experiments, empowering researchers to dynamically adjust experimental parameters based on immediate feedback from the analysis.
 
 Additionally, it incorporates uncertainty estimation in the fitted parameters using the MCMC (Monte Carlo Markov Chain) inverse algorithm, providing researchers with more reliable and nuanced results[@sunday2016mcmc].
 
@@ -72,11 +72,18 @@ This workflow ensures that the `cdsaxs` package not only identifies the optimal 
 
 # Comparision
 
-**Xi-cam**[@Xi-cam] is the open source software that was used for CD-SAXS simulations in this paper[@timothee]. Notably they use six trapezoid model and rounded trapezoid model to do the simulation. The experimental dataset used by Timothée et al.'s for their study was used here for a fit with six trapezoid model, enabling a comparison between Xi-cam and the `cdsaxs` package. The results are shown in the figure below:
[…]
