### 2025-01-26 87ca2a3
gpu info in paper and readme added
