### 2025-01-27 d02b7eb
Update paper.md
Rewrite a bit the comparison with Choisnet's work and the statement of need
