### 2025-01-27 7489c50
gaussian move vector mode activated and paper typo corrected
