### 2025-01-27 CEA-MetroCarac/cdsaxs#7 @mrhxszo
Thank you for the comment. Since most of the existing software are proprietary, we do not have access to them. However, there is one python code available on github in the package named Xi-cam. It is not maintained anymore but we were able to use the code in the current state.
We performed a comparison of results on a similar model based on a stacked of trapezoids. We started from identical profiles (with similar width, height and sidewal angles) and the results between the two models matched very closely. We've modified the paper to discuss about it and added a figure to illustrate the results.
