### 2025-01-27 CEA-MetroCarac/cdsaxs#6 @mrhxszo
We refined our text to provide more details on the previous work on cd-saxs modeling. We added the reference of studies conducted by NIST folks on the use of genetic algorithm and uncertainty calculation. We also modified the sentence in the text to give a clearer state of packages available for cd-saxs modeling.
