### 2025-01-27 #7559 @mrhxszo
@benjaminbolling I've answered all the issues. Can you check if it's up to your satisfaction?
@marcocamma I saw that in [this](https://github.com/CEA-MetroCarac/cdsaxs/issues/8) issue that you couldn't install the package my co-author has replied to your issue which might help you install it.
@srmnitc Can I as an author have the ability to complile the pdf version of the paper before a reviewer sees it ? If yes how?

Thank you everyone for your help.
