### 2025-01-27 #7559 @srmnitc
@mrhxszo Please take a look [here](https://joss.readthedocs.io/en/latest/paper.html#checking-that-your-paper-compiles). One of those should work!
