### 2025-02-18 CEA-MetroCarac/cdsaxs#15 @marcocamma
JOSS Review: Reproducibility and comparison
I like the paragraph about the comparison with other programs (notably Xi-cam).
Would it be possible to have include the experimental data and a script/notebook able to reproduce the cdsaxs fits shown in the "comparison" paragraph ?
Thanks !
