### 2025-02-21 CEA-MetroCarac/cdsaxs#7 @benjaminbolling
Resolved through [d02b7eb](https://github.com/CEA-MetroCarac/cdsaxs/commit/d02b7eb7ef57622369698dd9b614b33d5b4d0988)
