### 2025-02-21 CEA-MetroCarac/cdsaxs#6 @benjaminbolling
@mrhxszo, could you please describe where in the text (paper.md?) you refer to studies regarding the use of genetic algorithm calculation, conducted by NIST?
