### 2025-02-21 #7559 @danielskatz
@benjaminbolling - I believe you can do this, with the command `@editorialbot generate pdf`. editorialbot commands need to be the first entry in a new comment.
