### 2025-03-11 #7559 @mrhxszo
Thank you for your feedbacks i'll take a look at them in the coming week.
