### 2025-04-18 efc89d1
review 2 changes: paper, 2d, notebook compare
