+++ joss/compare/fit.py
@@ -1,0 +1,963 @@
+"""
+Performs model inverse resolution from cut inside QxQzI reciprocal map to
+obtained 3D line profile
+Based on trapezoïds line model and literal fourier transformed (extracted from
+XiCam an open source software
+
+Code developed by Jerome Reche and Vincent Gagneur.
+"""
+import os
+from collections import deque
+import numpy as np
+import pandas as pd
+import matplotlib.pyplot as plt
+import deap.base as dbase
+from deap import creator, tools, cma
+from scipy import stats
+import scipy.interpolate
+
+creator.create('FitnessMin', dbase.Fitness, weights=(-1.,))  # to minim. fitness
+creator.create('Individual', list, fitness=creator.FitnessMin)
+
+
+def find_peaks(values, positions, threshold):
+    """
+    Gives peaks over threshold as list of list of each point
+    TODO: check if this function already exist in numpy
+
+    Parameters
+    ----------
+    values: numpy.ndarray((n))
+        1D-data to analyse
+    positions: numpy.ndarray((n))
+        x-positions of data
+    threshold: float
+        threshold limit
+
+    Returns
+    -------
+    peaks: list of tuple of floats
+        list of peaks positions and related values
+    """
+    err_msg = "sizes of 'values' and 'positions' differ"
+    assert (len(values) == len(positions)), err_msg
+
+    first_point = 0
+    inside_peak = False
+    peaks = []
+    for (i, val) in enumerate(values):
+        # Detection of peak begining
+        if not inside_peak and val > threshold:
+            first_point = i
+            inside_peak = True
+        # Detection of end of peak
+        if inside_peak and (val < threshold or np.isnan(val)):
+            inside_peak = False
+            # At least four points must be used to have point
+            if (i - first_point) > 3:
+                peak = (positions[first_point - 1:i], values[first_point - 1:i])
+                peaks.append(peak)
+            elif i - first_point == 1:
+                peak = (positions[first_point], values[first_point])
+                peaks.append(peak)
+
+    return peaks
+
+
+def fit_peaks(peaks):
+    """
+    Return centers position of each peak and their  maximum
+    TODO: check why gauss fit was removed
+
+    Parameters
+    ----------
+    peaks: list of tuple of floats
+        list of peaks positions and related values
+
+    Returns
+    -------
+    centers_pos, max_peak: list of floats
+        All centers and values respectively obtained at this maximum
+    """
+    centers_pos = []
+    max_peak = []
+    for (pos, vals) in peaks:
+        if isinstance(pos, float):
+            centers_pos.append(pos)
+            max_peak.append(vals)
+        elif len(pos) > 1:
+            center, value = maxleftright(pos, vals)
+            centers_pos.append(center)
+            max_peak.append(value)
+
+    return centers_pos, max_peak
+
+
+def maxleftright(positions, values):
+    """
+    Return the center of a peak and the associated value by "search around"
+    TODO: check why gauss fit was removed
+
+    Parameters
+    ----------
+    positions: list of floats
+        Positions of data
+    values: list of floats
+        Values associated to positions
+
+    Returns
+    -------
+    center, value: floats
+        Center and associated value respectively
+    """
+    x0 = np.average(positions)
+
+    ind0 = np.where((positions >= x0 - 0.02) & (positions <= x0))
+    max0 = max(values[ind0])
+    pos0 = (positions[ind0])[np.where(values[ind0] == max0)]
+    ind1 = np.where((positions >= x0) & (positions <= x0 + 0.02))
+    max1 = max(values[ind1])
+    pos1 = (positions[ind1])[np.where(values[ind1] == max1)]
+
[…]
