### 2025-04-18 CEA-MetroCarac/cdsaxs#15 @mrhxszo
i've added a folder named "compare" inside the joss folder it contains the notebook and necessary files:
https://github.com/CEA-MetroCarac/cdsaxs/tree/main/joss/compare
