### 2025-04-18 CEA-MetroCarac/cdsaxs#5 @mrhxszo
I've changed it to:
By aiming to fill the current void in CD-SAXS data analysis tools, cdsaxs targets to speed up research workflows and to make advanced analytical techniques more accessible.
