### 2025-04-18 741ce8a
docstrings improved
