### 2025-04-18 492c805
updated docstrings
