### 2025-04-18 #7559 @mrhxszo
> * The "comparison paragraph" of the manuscript ends with "We can observe that the cdsaxs package has significant improvement in execution time (almost five times) over old code". From the graph, it seems that the speed-up is more than a factor 5.
> * please have a look at issue [JOSS Review: Reproducibility and comparison CEA-MetroCarac/cdsaxs#15](https://github.com/CEA-MetroCarac/cdsaxs/issues/15)
> * I am a bit surprised by the API (maybe just because I have not done these experiments myself)
>
>   * the `StackedTrapezoidSimulation` takes qys and qzs but it returns a 1D array
>   * I guess it is assumed that they have the same size (but no check is done when creating the instance) and that the intensity[i] is calculated for qys[i],qzs[i]
>   * The doc string could be a bit more explicit
>   * The code to get 2D images is not very elegant
>   * `x,y=np.meshgrid( np.linspace(-0.5,0.5,101), np.linspace(-0.5,0.5,101) )`
>   * `simulation = StackedTrapezoidSimulation(qys=x.ravel(), qzs=y.ravel())`
>   * `intensity = simulation.simulate_diffraction(params=i_params).reshape(x.shape)`
>   * again, maybe there is no need to get 2D arrays but otherwise I would suggest to change a bit the API

@marcocamma

Thank you very much for the review. I’ve addressed each point below:

   - I forgot to update the text to match the improved version, but everything is now consistent, please refer to the issue.

   - The notebook and all data needed to reproduce the results in the paper have been included.

   - Docstrings have been made more explicit, with detailed parameter descriptions. Further suggestions are welcome!

   - Based on user feedback, we determined that 2D profiles add little value—since our primary goal of model fitting isn’t possible in two dimensions—yet we’ve provided a simple option to enable 2D simulations when needed. You can now enable them via the `two_d` flag in `simulate_diffraction` as following.
   ```python
   qzs, qxs = np.meshgrid(
       np.linspace(-0.5, 0.5, 100),
       np.linspace(-0.5, 0.5, 200)
   )

   sim = StackedTrapezoidSimulation(qys=qxs, qzs=qzs)
   intensity_exp = sim.simulate_diffraction(params=sim_params, two_d=True)

```
