### 2025-04-18 #7559 @mrhxszo
> There are some references that are not rendered/part of the pdf generated albeit they are seen in the paper.md; namely _at{timothee}_ and _at{Xi-cam}_. I also noticed that albeit _at{sunday_2015}_ has an underscore in paper.md whilst it does not have an underscore (_at{sunday2015}_) in paper.bib, it is correctly rendered in the PDF.

Thanks for pointing that out. It should be okay now.
