### 2025-06-21 #7559 @benjaminbolling
I re-read the paper (sorry for the delay!), please find my comments in this task:
https://github.com/CEA-MetroCarac/cdsaxs/issues/16
