### 2025-06-21 #7559 @benjaminbolling
Left to do from my point of view:
- Quality of writing: Requires https://github.com/CEA-MetroCarac/cdsaxs/issues/16
- References: Requires https://github.com/CEA-MetroCarac/cdsaxs/issues/16
