### 2025-08-16 #7559 @benjaminbolling
Issues still found (sorry for nagging):
- Make sure there is spacing between a word and the invocation of a reference in the text
- You cannot enclose more than 1 reference per square bracket, separation needed with each reference object having their own square bracket
- Is it necessary for Xi-cam to be bold in the text?
