### 2025-08-29 #7559 @benjaminbolling
Thanks for the updates! Some remarks after looking through the PDF now:

1. Diagram illustrating the overall workflow of the CMAES algorithm should have a newline before the figure and figure label
2. The figure of the overall workflow of the MCMC algorithm should have a newline before the figure and figure label
3. Figure 3 label should be improved (execution time for the two versions of the code)
4. This is just a suggestion; add a conclusion section
