### 2026-01-23 #7559 @mrhxszo
A release has been made for the reviewed version of the software.

Version: v1.0.2
Zenodo DOI: https://doi.org/10.5281/zenodo.18347960

The archive metadata (title, authors, ORCIDs) and license match the JOSS paper and repository.
