### 2026-01-27 #7559 @srmnitc
@mrhxszo I opened an issue with some minor changes for the paper, could you please take a look?
