### 2026-02-08 #7559 @srmnitc
@mrhxszo One more quick thing before we wrap this up - could you please edit the Zenodo archive to say v1.0.2, instead of version - Joss_release ? Thanks!
