### 2026-02-09 #7559 @mrhxszo
i've changed the version name to  v1.0.2 instead of joss_relase for Zenodo
