### 2026-02-09 #7559 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
