### 2026-02-24 #7559 @rkurchin
👋 Hi, congrats on being nearly done! As AEiC for JOSS I will now help to process this submission for acceptance in JOSS (and apologies for the delay in getting this finalization process started!). Below are some final checks:
- [x] Copyedit pass/checked for typos
## Remaining points (all copyediting):
### Paper text
- [ ] I suggest spelling out the CEA acronym in the affiliation, as many international readers may not know what it stands for
- [x] Summary: Arguably, the miniaturization of transistors broadly is one of the major technological successes of the late 20th and early 21st century, so I might soften the phrasing here to something like "The further miniaturization of transistors, ..."
- [x] point 1 under Description: no need to capitalize Transform in Fourier transform
### Figure 1
- [x] I suggest adding a more descriptive caption that walks a reader through the figure a bit more (obviously you don't need to duplicate the text of the paper, but there are several things unexplained)
- [x] It looks like some kind of lossy compression or insufficient resolution of initial rendering, could you remake it so it's less fuzzy?
- [x] Also look out for a few places where I think a white box is partially blocking some text (e.g. the "g" in "best guess" or the "p" in "Experimental Data")
- [x] You've generally been using American English conventions in the text AFAICT, but you spell "optimise" the British way in the paper. As an international journal, JOSS doesn't mandate a particular convention, but you should be consistent throughout!
- [x] Re the equation in the upper right: I presume that's the model being fit, but this should be labeled/indicated somehow (or at least described in the figure caption), and also it looks like the image may have had its aspect ratio "squished" down a bit, so check on that
- [x] I suggest adding legends to the figures that have multiple datasets plotted (I assume it's measured and simulated, but not clear which is which)
### Figure 2
- [x] same suggestion as above about descriptive caption
- [x] Correct spelling is "Metropolis-Hastings"
- [x] please be consistent in plot titles about sentence vs. title case (upper left plot switches midway)
- [ ] Aesthetically, it looks as if the upper-right subfigure may be reproduced from elsewhere; if so, ensure you have permissions and that source should be cited
### Figure 3
- [x] again, this figure needs more explanation – what does each individual plot correspond to?
- [x] y-axis label should read "Intensity (A.U.)" (currently missing space and second period)
- [x] x-axis label should also have space before parenthetical units
### Figure 4
- [x] spaces between labels and units on both axes
- [x] add legend
