### 2026-03-24 #7559 @Kevin-Mattheus-Moerman
@mrhxszo can you confirm regarding this point raised before about figure 2:
_"Aesthetically, it looks as if the upper-right subfigure may be reproduced from elsewhere; if so, ensure you have permissions and that source should be cited"_
Do you have permission to reuse that subfigure here?
