### 2026-04-27 #7559 @mrhxszo
I have replaced the figure with this [Wikimedia Commons figure](https://commons.wikimedia.org/wiki/File:Flowchart-of-Metropolis-Hastings-M-H-algorithm-for-the-parameter-estimation-using-the.png#Licensing), which is distributed under a Creative Commons license, and I have updated the paper to cite both the original authors and Wikimedia.
