### 2026-05-02 #7559 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
